+++ src/Mesh.h
@@ -44,6 +44,6 @@ public:
         /*
          *Caculates the mixing length for all fluid objects 
          */
-    void calculateMixingLength(int dimX, int dimY, int dimZ, float dx, float dy, float dz, const vector<int> &icellflag, vector<float> &mixingLengths);
+    void calculateMixingLength(int dimX, int dimY, int dimZ, float dx, float dy, float dz, const vector<int> &icellflag, vector<double> &mixingLengths);
         
 };
