+++ src/URBGeneralData.h
@@ -103,7 +103,7 @@ public:
     std::vector<double> u,v,w;
 
 
-    std::vector<float> mixingLengths;
+    std::vector<double> mixingLengths;
 
     // Sensor* sensor;      may not need this now
